requirements.txt
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
pyannote.audio
pydub
librosa
torchaudio
tqdm
numpy
pandas
gdown
python-docx
requests
transformers
pandas 
yt-dlp 
boto3
minio
ipywidgets
jupyter
silero-vad
git+https://github.com/OpenPecha/fast-antx.git
